Move type annotation tests
