+++ test/expect/TestScriptPy3.test_string_print-stdout.expect
